+++ test/sandbox/app/components/turbo_stream_component.rb
@@ -1,0 +1,4 @@
+# frozen_string_literal: true
+
+class TurboStreamComponent < ViewComponent::Base
+end