+++ test/sandbox/app/controllers/integration_examples_controller.rb
@@ -70,5 +70,8 @@ class IntegrationExamplesController < ActionController::Base
 
   def unsafe_postamble_component
     render(UnsafePostambleComponent.new)
+
+  def turbo_stream
+    respond_to { |format| format.turbo_stream { render TurboStreamComponent.new } }
   end
 end